requirements.txt
# Data Manipulation & Analysis
pandas>=2.2.0
numpy>=1.26.0

# Machine Learning & Optimization
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
joblib>=1.3.0

# Handling Class Imbalance
imbalanced-learn>=0.12.0
boruta

# Explainability & Visualization
shap
matplotlib
seaborn
plotly

# Deep Learning
tensorflow

# Web Interface
streamlit>=1.50.0
pillow>=10.2.0

google.generativeai
xlsxwriter
